usa-name-trends/app/requirements.txt
streamlit==1.8.1
pandas==1.4.1
google-cloud-bigquery==2.34.2
matplotlib==3.5.1
altair==4.2.0